Remove the Method that Starts/Stops Connectivity Monitoring
Reason for change: Since Internet Monitoring is one of the core feature of NetworkManager Plugin and as lot other component listens to this event, This method will will be always triggered internally when the interface status changes. The external API to start/stop is removed.
Test Procedure: Check the connectivity Monitoring
Risks: Medium
Signed-off-by: Karunakaran A <karunakaran_amirthalingam@cable.comcast.com>